Code/AutoSage App Using Gemini Flash main/requirements.txt
streamlit
Pillow
google.generativeai
python-dotenv